Ongoing refactoring (use size_t instead of int)

+++ libtransmission/ConvertUTF.h
@@ -151,7 +151,7 @@ Boolean isLegalUTF8Sequence (const UTF8 *source, const UTF8 *sourceEnd);
 
 
 /* intended to work the same as g_utf8_validate */
-Boolean tr_utf8_validate (const char * str, int max_len, const char ** end);
+Boolean tr_utf8_validate (const char * str, size_t max_len, const char ** end);
 
 
 #ifdef __cplusplus
